[Docs] Remove legacy mobx example
