--- examples/with-mobx-react-lite/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["next/babel"],
-  "plugins": [["@babel/plugin-proposal-class-properties", { "loose": false }]]
-}
