--- examples/with-mobx-react-lite/package.json
@@ -1,19 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "mobx": "^6.0.3",
-    "mobx-react-lite": "^3.1.5",
-    "next": "latest",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
-  "devDependencies": {
-    "@babel/core": "7.14.5",
-    "@babel/plugin-proposal-class-properties": "^7.1.0"
-  }
-}
