--- examples/with-mobx-react-lite/pages/_app.js
@@ -1,9 +1,0 @@
-import { StoreProvider } from '../components/StoreProvider'
-
-export default function App({ Component, pageProps }) {
-  return (
-    <StoreProvider {...pageProps}>
-      <Component {...pageProps} />
-    </StoreProvider>
-  )
-}
