--- examples/with-mobx-react-lite/pages/index.js
@@ -1,5 +1,0 @@
-import Page from '../components/Page'
-
-export default function Index() {
-  return <Page title="Index Page" linkTo="/other" />
-}
