--- examples/with-mobx-react-lite/pages/other.js
@@ -1,5 +1,0 @@
-import Page from '../components/Page'
-
-export default function Other() {
-  return <Page title="Other Page" linkTo="/" />
-}
