--- examples/with-mobx/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": ["next/babel"],
-  "plugins": [
-    ["@babel/plugin-proposal-decorators", { "legacy": true }],
-    ["@babel/plugin-proposal-class-properties", { "loose": false }]
-  ]
-}
