+++ examples/with-mobx/components/Clock.js
@@ -1,8 +1,13 @@
-import { observer } from 'mobx-react'
-const Clock = observer((props) => {
+import { observer } from 'mobx-react-lite'
+import { useStore } from './StoreProvider'
+
+const Clock = observer(function Clock(props) {
+  // use store from the store context
+  const store = useStore()
+
   return (
-    <div className={props.light ? 'light' : ''}>
-      {props.timeString}
+    <div className={store.light ? 'light' : ''}>
+      {store.timeString}
       <style jsx>{`
         div {
           padding: 15px;
@@ -19,4 +24,5 @@ const Clock = observer((props) => {
     </div>
   )
 })
+
 export default Clock
