+++ examples/with-mobx/components/StoreProvider.js
