+++ examples/with-mobx/package.json
@@ -1,20 +1,15 @@
 {
   "private": true,
   "scripts": {
-    "dev": "next",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
-    "mobx": "^6.0.1",
-    "mobx-react": "^7.0.0",
+    "mobx": "^6.6.1",
+    "mobx-react-lite": "^3.4.0",
     "next": "latest",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
-  "devDependencies": {
-    "@babel/core": "7.14.5",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-proposal-decorators": "^7.1.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
